rename dir name
