+++ lex/src/Lexer.js
