+++ lex/src/defaults.js
