+++ lex/src/log.js
