+++ lex/src/rules.js
