+++ lex/src/token_test.js
@@ -5,11 +5,14 @@ const {Logger} = require(`./log.js`);
 
 const log = new Logger({shortTime: true});
 
-const text = `
-Email addresses in plain text are not linked: test@example.com.
-Email addresses wrapped in angle brackets are linked: <test@example.com>.
-They are also obfuscated so that email harvesting spam robots hopefully won not get them.
+const text = `| AA | BB | CC |
+| -- | -- | -- |
+| 1 | 2 | 3, **oo** | 
+| - 4, x | 5 | 6 *ss*  |
 `;
 const lex = new Lexer(defaults);
 const result = lex.lex(text);
-log.warn(result);
+
+const src = '```'+`
+
+`+'```';
